Fix `offer()` parameter signature

+++ src/d3a/models/market.py
@@ -88,7 +88,7 @@ class Market:
         for listener in sorted(self.notification_listeners, key=lambda l: random.random()):
             listener(event, market=self, **kwargs)
 
-    def offer(self, energy: int, price: int, seller: str) -> Offer:
+    def offer(self, energy: int, price: float, seller: str) -> Offer:
         if self.readonly:
             raise MarketReadOnlyException()
         if energy <= 0:
